refactor: extract SecretDetectionHandler class from executor.ts
